double fixed

+++ lib/page/bookKeeping/keeping_expenditure.dart
@@ -191,8 +191,8 @@ class _KeepExpenditureState extends State<KeepExpenditure> {
                 print(curImageNumString);
                 print(curCategoryName);
 
-                int _keepText =
-                    int.parse(_keepTextController.text) == 0
+                double _keepText =
+                    double.parse(_keepTextController.text) == 0
                         ? 0
                         : int.parse(_keepTextController.text);
                 String _beiZhuText = _beiZhuTextController.text == ''
